Finish _*ᴹ_
Maybe I have a new problem now; the proof term in the constructor is
hidden behind a lambda (to bring in the vectors for the inner product
proof), which means that I can't pattern match against the proof to use
the `refl` constructor. All the proofs should be the same due to the
Uniqueness of Identity Proofs (UIP).